requirements.txt
# Core
python-dotenv==1.0.0
loguru==0.7.2

# MongoDB
pymongo==4.6.1

# SQL
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Data Processing
pandas==2.2.0
numpy==1.26.3

# HTTP Client
requests==2.31.0
httpx==0.26.0

# Dashboard
streamlit==1.30.0
plotly==5.18.0
altair==5.2.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
tqdm==4.66.1
loguru==0.7.2
